requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1

requests==2.32.3
beautifulsoup4==4.12.3

sentence-transformers==3.0.1
faiss-cpu==1.8.0

numpy==1.26.4

# Optional (only if you decide to use Selenium later)
# selenium==4.23.1
# webdriver-manager==4.0.2
